add conftest.py to hold common fixtures
(feature/pytest)
