+++ tests/test_config.py
@@ -1,26 +1,12 @@
 # -*- coding: utf-8 -*-
 import click
-import git
 import os
 import pytest
 import yaml
-import tempfile
 import zazu.config
 
 __author__ = "Nicholas Wiles"
 __copyright__ = "Copyright 2016"
-
-
-@pytest.fixture()
-def git_repo():
-    dir = tempfile.mkdtemp()
-    repo = git.Repo.init(dir)
-    readme = os.path.join(dir, 'README.md')
-    with open(readme, 'w'):
-        pass
-    repo.index.add([readme])
-    repo.index.commit('initial readme')
-    return repo
 
 
 @pytest.fixture()
